# core/pool-domain-manager/api/openapi.yaml
openapi: 3.0.3
info:
  title: pool-domain-manager
  version: 1.0.0
  description: |
    Central authority for OpenOva-pool subdomain allocation. Closes #163.
    The PDM is the SOLE source of truth for which (poolDomain, subdomain)
    pairs have been reserved or activated across the OpenOva fleet, and the
    SOLE service in the fleet that calls api.dynadot.com.

    State machine per (domain, subdomain) pair:

        NULL ─reserve→ RESERVED ─commit→ ACTIVE
                          │              │
                          expire/        release/
                          release        destroy
                          ↓              ↓
                        NULL          NULL

servers:
  - url: http://pool-domain-manager.openova-system.svc.cluster.local:8080
    description: In-cluster catalyst-api → PDM call path
  - url: https://pool.openova.io
    description: Operator-facing endpoint (auth-gated)

paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: PDM is healthy and CNPG is reachable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  managedDomains:
                    type: array
                    items: { type: string }
        '503':
          description: PDM is up but CNPG is unreachable

  /api/v1/reserved:
    get:
      summary: Canonical reserved-subdomain list
      responses:
        '200':
          description: List of reserved subdomain labels
          content:
            application/json:
              schema:
                type: object
                properties:
                  reserved:
                    type: array
                    items: { type: string }

  /api/v1/pool/{domain}/check:
    get:
      summary: Fast availability read
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
        - in: query
          name: sub
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Always 200 — clients use body.available, not status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResult'

  /api/v1/pool/{domain}/reserve:
    post:
      summary: Atomic reserve with TTL
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subdomain]
              properties:
                subdomain: { type: string }
                createdBy: { type: string }
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReserveResponse'
        '409':
          description: Subdomain already taken
        '422':
          description: Invalid input (format / unsupported pool)

  /api/v1/pool/{domain}/commit:
    post:
      summary: Promote reservation → active and write Dynadot records
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subdomain, reservationToken, sovereignFQDN, loadBalancerIP]
              properties:
                subdomain: { type: string }
                reservationToken: { type: string, format: uuid }
                sovereignFQDN: { type: string }
                loadBalancerIP: { type: string, format: ipv4 }
                # #4053 — optional. When set, the console + api A-records point
                # at the dedicated console LB (the isolated cilium-gateway-console)
                # instead of loadBalancerIP, so a poisoned shared-gateway CEC can
                # never 404 the console. Omitted = every record points at
                # loadBalancerIP (legacy single-LB behaviour).
                consoleLoadBalancerIP: { type: string, format: ipv4 }
      responses:
        '200': { description: Committed }
        '202': { description: Committed in DB; Dynadot retry needed }
        '403': { description: Reservation token mismatch }
        '404': { description: No reservation exists }
        '409': { description: Already active }
        '410': { description: Reservation TTL expired }

  /api/v1/pool/{domain}/release:
    delete:
      summary: Free a (pool, subdomain) and remove Dynadot records
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [subdomain]
              properties:
                subdomain: { type: string }
      responses:
        '200': { description: Released }
        '202': { description: Row deleted; Dynadot delete partial }
        '404': { description: No allocation exists }

  /api/v1/pool/{domain}/list:
    get:
      summary: Operator-facing list of allocations
      parameters:
        - in: path
          name: domain
          required: true
          schema: { type: string }
      responses:
        '200':
          description: All allocations for the pool

  /api/v1/registrar/{registrar}/set-dns:
    post:
      summary: Write raw zone-level DNS records (A/AAAA/CNAME/MX/TXT) via the registrar's record-level API
      description: |
        REPLACES the record set for the domain. Records NOT in the new set are
        DELETED; records in the new set are CREATED or UPDATED. Implementations
        MUST be idempotent.

        Adapters that lack a record-level API return 501. Use `/set-ns` to
        delegate DNS to a record-capable provider (PowerDNS, Route53, etc.)
        instead.

        Refs chepherd/chepherd#36.
      parameters:
        - in: path
          name: registrar
          required: true
          schema: { type: string }
          description: Lowercase adapter name (currently `dynadot` or `namecheap`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain, token, records]
              properties:
                domain:
                  type: string
                  example: example.com
                token:
                  type: string
                  description: Registrar API credentials. Format varies by adapter (Dynadot = `key:secret`, Namecheap = `apiUser:apiKey:user:clientIP`). NEVER logged.
                records:
                  type: array
                  minItems: 1
                  items: { $ref: '#/components/schemas/DNSRecord' }
      responses:
        '200':
          description: Records written + read-back confirmation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, enum: [true] }
                  registrar: { type: string }
                  domain: { type: string }
                  records:
                    type: array
                    items: { $ref: '#/components/schemas/DNSRecord' }
                  propagation:
                    type: string
                    description: Coarse human-readable estimate (e.g. "up to 24 hours; typically 1-4 hours")
        '401': { description: Invalid token }
        '403': { description: Domain not in account }
        '404': { description: Unknown registrar }
        '422': { description: Invalid records (type not in A/AAAA/CNAME/MX/TXT, empty value, etc.) }
        '429': { description: Registrar rate-limited }
        '501': { description: Registrar adapter does not implement DNSRegistrar — use /set-ns to delegate DNS instead }
        '502': { description: Registrar API unavailable }

components:
  schemas:
    DNSRecord:
      type: object
      required: [type, value]
      properties:
        subhost:
          type: string
          description: Leftmost label(s). Empty = root of zone. `www` = `www.<domain>`.
          example: www
        type:
          type: string
          enum: [A, AAAA, CNAME, MX, TXT]
        value:
          type: string
          example: 1.2.3.4
        priority:
          type: integer
          minimum: 0
          description: MX priority (0 = highest). Ignored for non-MX records.
        ttl:
          type: integer
          minimum: 0
          description: Seconds. 0 = use registrar's default (typically 3600).

    CheckResult:
      type: object
      properties:
        available: { type: boolean }
        reason: { type: string }
        detail: { type: string }
        fqdn: { type: string }

    ReserveResponse:
      type: object
      properties:
        poolDomain: { type: string }
        subdomain: { type: string }
        state: { type: string, enum: [reserved] }
        reservedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        reservationToken: { type: string, format: uuid }
        createdBy: { type: string }